Factor our type type checking into a separate file. No functionality change
