+++ frontends/CMakeLists.txt
@@ -71,6 +71,7 @@ set (P4_FRONTEND_SRCS
   p4/typeChecking/bindVariables.cpp
   p4/typeChecking/syntacticEquivalence.cpp
   p4/typeChecking/typeChecker.cpp
+  p4/typeChecking/typeCheckTypes.cpp
   p4/typeChecking/typeConstraints.cpp
   p4/typeChecking/typeSubstitution.cpp
   p4/typeChecking/typeSubstitutionVisitor.cpp
